Lesson22 hw develop->main

+++ Lesson21/src/main/java/by/tms/utils/CRUDUtils.java
@@ -10,21 +10,21 @@ import static by.tms.utils.DBUtils.getConnection;
 
 @UtilityClass
 public class CRUDUtils {
-    public static void updateOneParameterById(Integer id, String parameter, String updateQuery) {
+    public static void updateOneParameterById(Long id, String parameter, String updateQuery) {
         try (Connection connection = getConnection()) {
             PreparedStatement statement = connection.prepareStatement(updateQuery);
             statement.setString(1, parameter);
-            statement.setInt(2, id);
+            statement.setLong(2, id);
             statement.executeUpdate();
         } catch (SQLException e) {
             System.out.println("Exception: " + e.getMessage());
         }
     }
 
-    public static void deleteById(Integer id, String deleteQuery) {
+    public static void deleteById(Long id, String deleteQuery) {
         try (Connection connection = getConnection()) {
             PreparedStatement statement = connection.prepareStatement(deleteQuery);
-            statement.setInt(1, id);
+            statement.setLong(1, id);
             statement.execute();
         } catch (SQLException e) {
             System.out.println("Exception: " + e.getMessage());
